% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code.R
\name{lbs_to_kgs}
\alias{lbs_to_kgs}
\title{Convert weight of object from Pounds to Kilograms}
\usage{
lbs_to_kgs(weight_in_lbs)
}
\arguments{
\item{x}{Weight of object in pounds which is to be converted in kilograms}
}
\value{
Weight of object in Kilograms
}
\description{
Takes weight of any object in pounds and convert it into Kilograms
}
